refactor(wms): cut count over to inventory_adjustment and retire legacy count tree
