--- app/wms/count/contracts/__init__.py
@@ -1,8 +1,0 @@
-from __future__ import annotations
-
-from .count import CountRequest, CountResponse
-
-__all__ = [
-    "CountRequest",
-    "CountResponse",
-]
