--- app/wms/count/repos/__init__.py
@@ -1,7 +1,0 @@
-from __future__ import annotations
-
-from .count_repo import CountRepo
-
-__all__ = [
-    "CountRepo",
-]
