--- app/wms/count/routers/__init__.py
@@ -1,7 +1,0 @@
-from __future__ import annotations
-
-from .count import router
-
-__all__ = [
-    "router",
-]
