--- app/wms/count/services/__init__.py
@@ -1,7 +1,0 @@
-from __future__ import annotations
-
-from .count_service import CountService
-
-__all__ = [
-    "CountService",
-]
